[refactor] move code for calculating engagement summary to engagement model
Originally, the summary stat for an engagement is calculated within
iteration_controller. We decided that calculating the stat is the
engagement model's job, so we moved the code. In addition to moving
the code to model, more refactorings were done in iteration model.
Some changes were applied to index view and stylesheets as well.