[refactor] reduce cyclomatic complexity for creation_controller#create
Since our models are associated (user <- org <- app), the order in which
we create an object is important. Also, if any of the object creation fail
then no object should be created. That is, a transaction has to fail
and roll back if needed. Instead of doing the rollback using if statements
use ActiveRecord::Base.transaction that deals with this problem.